==> 03_subset_rospravosudie_to_homicides.R <==
library(data.table)  # Импорт библиотеки data.table для работы с таблицами

### Rospravosudie
verdicts <- fread("rps_dataset.csv")  # Чтение данных из CSV файла в таблицу данных (data.table)

# Фильтрация для простых убийств
hom <- verdicts[category %in% c("105 ч.1", "111 ч.4")]  # Фильтрация строк, где категория дела соответствует "105 ч.1" или "111 ч.4" (примеры статей Уголовного кодекса РФ)

# Фильтрация для приговоров в суде первой инстанции
hom <- hom[vid_dokumenta == "приговор" & etapd == "Первая инстанция"][, 
                                                                      year := as.numeric(
                                                                        substr(path, 
                                                                               nchar(path) - 20, 
                                                                               nchar(path) - 17)
                                                                      )
]  # Фильтрация строк, где тип документа - "приговор" и этап дела - "Первая инстанция". Добавление колонки "year", извлекая год из пути к файлу.

hom <- hom[year %in% c(2011)]  # Фильтрация строк, где год равен 2011

hom[region == "Читинская область", region := "Забайкальский край"]  # Замена названия региона "Читинская область" на "Забайкальский край"

# Удаление строк для региона Байконур
hom <- hom[region != ""]  # Удаление строк, где регион пустой (для региона Байконур)

fwrite(hom, "C:/Users/WarSa/OneDrive/Рабочий стол/диплом/rospravosudie_sou/rospravosudie_homicides_2011.csv")  # Сохранение отфильтрованных данных в новый CSV файл
