==> 06_move_marked_up.py <==
import os
import shutil

html_dir = "C:/Users/WarSa/OneDrive/Рабочий стол/диплом/rospravosudie_sou/verdicts/"
new_directory = (
    "C:/Users/WarSa/OneDrive/Рабочий стол/диплом/rospravosudie_sou/marked_verdicts/"
)


lst = [
    1017,
    8004,
    4775,
    9725,
    8462,
    4050,
    8789,
    1301,
    8522,
    1799,
    9941,
    8229,
    1129,
    8921,
    878,
    8677,
    7845,
    5922,
    6526,
    5071,
    2900,
    7075,
    4650,
    9638,
    2159,
    3476,
    1948,
    2580,
    1530,
    7289,
    4633,
    6519,
    4344,
    1222,
    2858,
    5400,
    9888,
    526,
    1069,
    5522,
    8214,
    5838,
    5862,
    6337,
    5491,
    9320,
    7494,
    9175,
    1791,
    9831,
    4939,
    465,
    7976,
    9392,
    3863,
    9326,
    8276,
    1895,
    3101,
    3990,
    6877,
    6764,
    1328,
    8749,
    8479,
    6473,
    6995,
    9714,
    858,
    6763,
    9182,
    316,
    733,
    6273,
    4907,
    5051,
    2330,
    9359,
    6429,
    3992,
    1706,
    501,
    6655,
    536,
    6881,
    7507,
    7514,
    3747,
    5163,
    29,
    1942,
    4934,
    1820,
    2281,
    9509,
    1966,
    8561,
    4182,
    5603,
    5742,
    8547,
    5939,
    6956,
    9265,
    8553,
    9532,
    9458,
    6902,
    2866,
    8943,
    219,
    8327,
    532,
    3123,
    9104,
    6611,
    2178,
    5936,
    7233,
    5932,
    4455,
    2153,
    1148,
    1101,
    1242,
    1218,
    4115,
    8465,
    418,
    6563,
    867,
    4499,
    3821,
    9803,
    5897,
    818,
    4730,
    664,
    719,
    500,
    3045,
    6816,
    6373,
    8615,
    8159,
    8613,
    9181,
    3598,
    3700,
    5136,
    9346,
    8696,
    785,
    6789,
    7131,
    8053,
    1572,
    4401,
    1833,
    2461,
    4225,
    309,
    7892,
    4078,
    7146,
    8633,
    3189,
    470,
    1360,
    1822,
    1790,
    8541,
    8066,
    3144,
    9086,
    4686,
    8666,
    9943,
    5048,
    455,
    6500,
    7793,
    3306,
    6217,
    8207,
    4158,
    4486,
    9860,
    9251,
    5987,
    10076,
    5447,
    6809,
    8174,
    4096,
    6887,
    4084,
    2110,
    1172,
    797,
    10054,
    9788,
    4072,
    5057,
    7452,
    5388,
    2762,
    8802,
    1265,
    7812,
    9991,
    2265,
    2156,
    6415,
    5390,
    9979,
    7151,
    5458,
    5729,
    1760,
    5704,
    5685,
    4664,
    4075,
    5474,
    7525,
    4212,
    3070,
    5353,
    7439,
    9655,
    9413,
    462,
    6443,
    4049,
    6657,
    9690,
    9852,
    7888,
    2033,
    731,
    7039,
    2813,
    3805,
    4771,
    966,
    9383,
    6193,
    5591,
    5793,
    8483,
    4749,
    8439,
    2482,
    6885,
    3388,
    3418,
    2722,
    2884,
    56,
    7193,
    934,
    7640,
    8109,
    4590,
    7883,
    3552,
    5123,
    6803,
    4257,
    127,
    717,
    9644,
    7715,
    7447,
    676,
    10057,
    5795,
    148,
    121,
    8361,
    8790,
    1491,
    2815,
    9772,
    1115,
    6308,
    3616,
    3551,
    6525,
    436,
    2137,
    8137,
    533,
    9573,
    2752,
    5592,
    9038,
    5191,
    6459,
    8411,
    9050,
    3565,
    570,
    6976,
    1926,
    5666,
    1813,
    7681,
    2833,
    7885,
    5694,
    7981,
    9851,
    5161,
    7560,
    8473,
    4063,
    7918,
    9952,
    724,
    1354,
    8684,
    6164,
    3255,
    9866,
    5199,
    8706,
    6017,
    9445,
    2016,
    132,
    6186,
    5094,
    1659,
    6068,
    6438,
    9613,
    2312,
    6321,
    5040,
    1839,
    3505,
    1838,
    5259,
    9736,
    8512,
    4211,
    9666,
    6595,
    9434,
    6990,
    6303,
    2167,
    2543,
    5914,
    8301,
    3760,
    6663,
    2587,
    8945,
    10014,
    7935,
    393,
    8840,
    7046,
    7923,
    4208,
    5147,
    4318,
    6410,
    4254,
    9833,
    2778,
    9758,
    7343,
    6103,
    7334,
    9331,
    4436,
    6236,
    5772,
    99,
    7684,
    7983,
    3200,
    3822,
    8090,
    6989,
    5763,
    7805,
    1190,
    1459,
    5117,
    1776,
    116,
    7360,
    9701,
    3157,
    1749,
    1326,
    1847,
    9424,
    1018,
    9946,
    383,
    8920,
    771,
    5207,
    484,
    4599,
    6449,
    3887,
    2563,
    58,
    4801,
    2021,
    9823,
    2170,
    1803,
    9606,
    1124,
    313,
    6790,
    822,
    6625,
    5480,
    5131,
    3382,
    9775,
    81,
    3328,
    3126,
    2194,
    5235,
    5663,
    2796,
    6472,
    2775,
    7174,
    1401,
    8627,
    5605,
    7961,
    3360,
    8426,
    1779,
    7656,
    4278,
    8253,
    4606,
    1089,
    8841,
    1861,
    9978,
    7374,
    7893,
    9650,
    2992,
    3797,
    6079,
    1177,
    8209,
    1057,
    3706,
    6440,
    7123,
    2257,
    1525,
    6420,
    5809,
    2079,
    9935,
    8987,
    7009,
    8765,
    3037,
    6485,
    3558,
    4568,
    7842,
    2351,
    7666,
    3651,
    6266,
    1871,
    5034,
    5005,
    8060,
    363,
    8816,
    2573,
    9459,
    9730,
    3329,
    2373,
    702,
    1378,
    959,
    2273,
    9962,
    8589,
    5031,
    9878,
    744,
    564,
    9883,
    2290,
    3482,
    5327,
    5287,
    87,
    4313,
    5633,
    860,
    5620,
    7817,
    6134,
    2446,
    628,
    5368,
    1104,
    1493,
    4062,
    3926,
    4784,
    2811,
    1473,
    2572,
    4364,
    9627,
    8863,
    4933,
    2569,
    9410,
    4745,
    2551,
    128,
    7211,
    1435,
    102,
    7255,
    4066,
    3457,
    453,
    4585,
    6138,
    9321,
    8146,
    6050,
    815,
    8766,
    644,
    7238,
    8240,
    8151,
    3437,
    7120,
    9366,
    2608,
    4358,
    1590,
    7368,
    3352,
    9027,
    2899,
    9821,
    1821,
    9751,
    7917,
    8793,
    1759,
    8910,
    2107,
    1019,
    2020,
    8735,
    7757,
    9675,
    296,
    5144,
    5853,
    764,
    5566,
    8314,
    3047,
    8690,
    1043,
    376,
    4562,
    6101,
    919,
    9964,
    1487,
    6109,
    5999,
    5661,
    7025,
    854,
    8717,
    10042,
    108,
    10053,
    3862,
    3018,
    2285,
    8588,
    144,
    3733,
    5254,
    9826,
    8444,
    2069,
    4963,
    133,
    9879,
    386,
    1802,
    8482,
    8962,
    5756,
    7500,
    9094,
    2158,
    4297,
    487,
    7379,
    3254,
    1421,
    908,
    3363,
    247,
    8054,
    8649,
    9183,
    6320,
    8775,
    647,
    2732,
    492,
    3807,
    3586,
    1620,
    7348,
    3129,
    5435,
    553,
    1864,
    1443,
    5238,
    7445,
    7200,
    1031,
    4330,
    2531,
    3417,
    9529,
    1587,
    9824,
    1651,
    3402,
    6818,
    5776,
    5195,
    7512,
    1630,
    7998,
    8864,
    5527,
    2136,
    9559,
    5148,
    8343,
    1524,
    8978,
    3946,
    827,
    4876,
    2616,
    2684,
    8825,
    6167,
    8968,
    7419,
    1131,
    4333,
    1890,
    9395,
    5334,
    4404,
    2997,
    7637,
    6398,
    9934,
    8378,
    4407,
    2019,
    2548,
    5750,
    1967,
    4279,
    9752,
    8539,
    4490,
    6557,
    8781,
    5305,
    6583,
    2161,
    5630,
    4210,
    8692,
    1216,
    2395,
    8423,
    129,
    2424,
    3145,
    2792,
    314,
    4324,
    8881,
    8101,
    823,
    8164,
    407,
    5024,
    9828,
    280,
    3579,
    517,
    3793,
    3590,
    8861,
    851,
    10002,
    3465,
    9449,
    6314,
    7652,
    793,
    703,
    324,
    8427,
    3641,
    3390,
    2106,
    787,
    8068,
    1373,
    1592,
    660,
    5006,
    1038,
    6574,
    9487,
    64,
    2131,
    9112,
    626,
    8498,
    6850,
    5607,
    5502,
    4848,
    2506,
    6293,
    8097,
    8567,
    5920,
    284,
    6940,
    1863,
    1950,
    8094,
    8608,
    8168,
    3210,
    4759,
    4973,
    9300,
    4023,
    9864,
    10023,
    7173,
    7555,
    10012,
    9794,
    4206,
    3729,
    6849,
    4400,
    5512,
    7891,
    384,
    6776,
    5330,
    9972,
    5846,
    5943,
    9324,
    8644,
    573,
    1611,
    5165,
    8345,
    1700,
    3781,
    479,
    1922,
    7914,
    8544,
    8389,
    4083,
    3298,
    9992,
    5863,
    4186,
    4064,
    5671,
    2664,
    8971,
    8827,
    249,
    9827,
    7239,
    101,
    7965,
    3665,
    520,
    5670,
    8313,
    9421,
    3415,
    7435,
    650,
    617,
    4676,
    1846,
    4223,
    3680,
    7568,
    1428,
    7300,
    4105,
    4944,
    5361,
    3338,
    5864,
    9101,
    9309,
    5337,
    4754,
    8778,
    812,
    2480,
    5956,
    338,
    7809,
    5845,
    6512,
    8398,
    8634,
    4777,
    8147,
    2267,
    6919,
    8782,
    9677,
    7310,
    1372,
    6063,
    6081,
    9009,
    2823,
    220,
    7095,
    6558,
    5413,
    6584,
    4572,
    2214,
    4637,
    8744,
    3663,
    6581,
    8714,
    611,
    7671,
    217,
    9371,
    8654,
    4819,
    6239,
    7309,
    3631,
    6515,
    6531,
    7826,
    271,
    1299,
    9895,
    2871,
    6799,
    7821,
    4171,
    7304,
    5957,
    4893,
    4733,
    27,
    8753,
    6448,
    4578,
    4385,
    9247,
    2937,
    4292,
    3766,
    7228,
    7261,
    9927,
    8443,
    5061,
    9080,
    3115,
    2223,
    5912,
    9480,
    1598,
    8405,
    5575,
    5934,
    8693,
    1088,
    4721,
    2761,
    9051,
    3572,
    1071,
    6348,
    9904,
    503,
    2541,
    5466,
    2439,
    5857,
    2864,
    979,
    3152,
    9290,
    2284,
    610,
    9242,
    8682,
    7574,
    6226,
    2116,
    656,
    10086,
    4327,
    3702,
    6693,
    5135,
    5176,
    7486,
    1252,
    2826,
    9407,
    2391,
    2678,
    9400,
    413,
    8395,
    8671,
    6660,
    7785,
    7987,
    8331,
    1259,
    5189,
    7740,
    4901,
    4103,
    9762,
    4895,
    7799,
    1654,
    427,
    9055,
    9357,
    8131,
    582,
    8768,
    8045,
    1817,
    225,
    6247,
    5578,
    8377,
    2841,
    2426,
    4609,
    5770,
    8437,
    3679,
    3842,
    2928,
]

if not os.path.exists(new_directory):
    os.makedirs(new_directory)

for i in range(200):
    file_to_move = os.path.join(html_dir, f"{lst[i]}.html")
    new_file_path = os.path.join(new_directory, f"{lst[i]}.html")
    shutil.move(file_to_move, new_file_path)
